+++ src/com/insomniacmath/Utils.java
@@ -3,11 +3,11 @@ package com.insomniacmath;
 
 public class Utils {
 
-    public  static String round(float i) {
+    public  static String round(double i) {
         if (i % 1 == 0)
             return Integer.toString((int) i);
         else
-            return Float.toString(i);
+            return Double.toString(i);
     }
 
 }
